Move types into own module
